Fix fast refresh compatibility for income statements
