+++ frontend/src/citizen-frontend/income-statements/IncomeStatementComponents.tsx
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import React, { useCallback } from 'react'
+import React from 'react'
 import styled, { useTheme } from 'styled-components'
 
 import { tabletMin } from 'lib-components/breakpoints'
@@ -12,26 +12,6 @@ import { FixedSpaceRow } from 'lib-components/layout/flex-helpers'
 import { fontWeights, Label } from 'lib-components/typography'
 import { defaultMargins, Gap } from 'lib-components/white-space'
 import { fasExclamationTriangle } from 'lib-icons'
-
-export type SetStateCallback<T> = (fn: (prev: T) => T) => void
-
-export function useFieldSetState<T, K extends keyof T>(
-  onChange: SetStateCallback<T>,
-  key: K
-): SetStateCallback<T[K]> {
-  return useCallback<SetStateCallback<T[K]>>(
-    (fn) => onChange((prev) => ({ ...prev, [key]: fn(prev[key]) })),
-    [onChange, key]
-  )
-}
-
-export function useFieldDispatch<T, K extends keyof T>(
-  onChange: SetStateCallback<T>,
-  key: K
-): (value: T[K]) => void {
-  const setState = useFieldSetState(onChange, key)
-  return useCallback((value: T[K]) => setState(() => value), [setState])
-}
 
 export const LabelWithError = React.memo(function LabelWithError({
   label,
@@ -49,10 +29,6 @@ export const LabelWithError = React.memo(function LabelWithError({
     </FixedSpaceRow>
   )
 })
-
-export function identity<T>(value: T): T {
-  return value
-}
 
 export interface IncomeStatementFormAPI {
   scrollToErrors: () => void
